Tag version 3.2.2

+++ src/gnuwin32/system.c
@@ -15,7 +15,7 @@
  *
  *  You should have received a copy of the GNU General Public License
  *  along with this program; if not, a copy is available at
- *  https://www.R-project.org/Licenses/
+ *  http://www.r-project.org/Licenses/
  */
 
 /* See ../unix/system.txt for a description of functions */
@@ -57,6 +57,7 @@ void editorcleanall(void);                  /* from editor.c */
 int Rwin_graphicsx = -25, Rwin_graphicsy = 0;
 
 R_size_t R_max_memory = INT_MAX;
+Rboolean UseInternet2 = TRUE; // used in main/internet.c
 
 extern SA_TYPE SaveAction; /* from ../main/startup.c */
 Rboolean DebugMenuitem = FALSE;  /* exported for rui.c */
@@ -493,16 +494,8 @@ void R_CleanUp(SA_TYPE saveact, int status, int runLast)
 	PrintWarnings();        /* from device close and (if run) .Last */
     app_cleanup();
     RConsole = NULL;
-    // Add some protection against calling this more than once:
-    // caused by signals on Unix, so maybe cannot happen here.
-    if(ifp) { 
-	fclose(ifp);    /* input file from -f or --file= */
-	ifp = NULL; 
-    }
-    if(ifile[0]) {
-	unlink(ifile); /* input file from -e */
-	ifile[0] = NULL;
-    }
+    if(ifp) fclose(ifp);        /* input file from -f or --file= */
+    if(ifile[0]) unlink(ifile); /* input file from -e */
     exit(status);
 }
 
@@ -990,7 +983,7 @@ int cmdlineoptions(int ac, char **av)
 		InThreadReadConsole = FileReadConsole;
 		setvbuf(stdout, NULL, _IONBF, 0);
 	    } else if (!strcmp(*av, "--internet2")) {
-/*	        This is now the default */
+/*		UseInternet2 = TRUE;    This is now the default */
 	    } else if (!strcmp(*av, "--mdi")) {
 		MDIset = 1;
 	    } else if (!strcmp(*av, "--sdi") || !strcmp(*av, "--no-mdi")) {
